# content/posts/k8s-learn-2-kubernetes_api_database_integration/ignore_tutorial_files/swagger.yaml
openapi: 3.0.3
info:
  title: Messages API
  description: A simple API for managing messages.
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /api/messages:
    get:
      summary: Get all messages
      description: Retrieve a list of all messages in the database.
      responses:
        "200":
          description: A JSON array of messages.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: The message ID.
                    message:
                      type: string
                      description: The message content.
        "500":
          description: Error connecting to the database or fetching messages.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: An error message.
    post:
      summary: Add a new message
      description: Add a new message to the database.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  description: The message content.
      responses:
        "201":
          description: Message added successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: A success message.
        "400":
          description: Message is required.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: An error message.
        "500":
          description: Error connecting to the database or adding a message.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: An error message.